Create README.md
Our first step is to reduce the dimension of genes to find sets of genes which are most likely to
predict the response. To achieve this, we use three methods which are joint mutual information,
Pearson Correlation and Wilcoxson Rank-Sum Test. Specifically, we want to leave 20 most
informative genes, so we find three sets of genes A, B and C among 13299 genes. Genes in A have
20 highest joint mutual information of genes and the response. Genes in B have 20 highest
Pearson Correlation to the response. Genes in C have 20 smallest P-values of Wilcoxson RankSum test with the null hypothesis values of a gene related to two labels are equally spread. In the second step, we split our data into training and test set with proportion 8:2. Notice that our data is
greatly imbalanced, we keep the proportion same for two labels of response in training and test
set. In addition, we can also implement weighted methods to eliminate the effect of imbalance.
Next step is to use different machine learning methods, particularly, Logistics Regression, SVM,
LDA, QDA, Random Forest and Artificial Neural Network, to train and predict using A, B, C
three sets. We try to find the best methods and gene set based on ROC curve and AUC (Area
under curve), since the data is imbalanced AUC instead of accuracy score is an appropriate
criterion for the prediction.